refactor: migrate deck start page
