+++ src/features/deck/index.ts
@@ -1,4 +1,6 @@
 export { DeckForm, type DeckFormFields, type DeckFormProps } from "./components/DeckForm";
+export { DeckStartForm, type DeckStartFormProps } from "./components/DeckStartForm";
 export { useDeckActions } from "./hooks/useDeckActions";
+export { useDeckFilterState } from "./hooks/useDeckFilterState";
 export { useDeckMutations } from "./hooks/useDeckMutations";
 export { deckFormSchema, type DeckFormValues } from "./lib/deckFormSchema";
