+++ src/features/study/index.ts
@@ -1,4 +1,3 @@
-export { DeckStartContainer } from "./containers/DeckStartContainer";
 export { DeckSwiperContainer } from "./containers/DeckSwiperContainer";
 export {
   discardStudySessionsMissingDecks,
@@ -6,5 +5,6 @@ export {
   touchStudySession,
 } from "./commands/studySessionCommands";
 export { useStudyHydrated } from "./hooks/useStudyHydrated";
+export { useStudyActions } from "./hooks/useStudyActions";
 export { useStudySessions } from "./hooks/useStudySessions";
 export type { StudySession } from "./state/studyStore";
